Refactored AttributesObject into shared Attributes module.
